# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vehicles": {
      "get": {
        "operationId": "VehiclesController_findAll",
        "summary": "Liste des véhicules (filtrée selon le rôle)",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "AVAILABLE",
                "ASSIGNED",
                "IN_SERVICE",
                "IMMOBILIZED",
                "IN_REPAIR",
                "ACCIDENTED",
                "PENDING_INSPECTION",
                "REPOSSESSED",
                "OUT_OF_SERVICE",
                "SOLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "managerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "VehiclesController_create",
        "summary": "Créer un véhicule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findById",
        "summary": "Détail véhicule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update",
        "summary": "Modifier un véhicule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove",
        "summary": "Supprimer (soft delete) un véhicule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/vehicles/{id}/assign-manager": {
      "post": {
        "operationId": "VehiclesController_assignManager",
        "summary": "Affecter un manager au véhicule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignManagerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/vehicles/{id}/status": {
      "patch": {
        "operationId": "VehiclesController_updateStatus",
        "summary": "Changer le statut du véhicule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "vehicles"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet Platform API",
    "description": "API de gestion de flotte VTC — Fleet Platform V2",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "plateNumber": {
            "type": "string",
            "example": "DK 1234 AB"
          },
          "brand": {
            "type": "string",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "example": "HiAce"
          },
          "vin": {
            "type": "string",
            "example": "AA123456789012345"
          },
          "year": {
            "type": "number",
            "example": 2022
          },
          "color": {
            "type": "string",
            "example": "Blanc"
          },
          "fuelType": {
            "type": "string",
            "example": "DIESEL",
            "enum": [
              "GASOLINE",
              "DIESEL",
              "HYBRID",
              "ELECTRIC"
            ]
          },
          "transmission": {
            "type": "string",
            "example": "MANUAL",
            "enum": [
              "MANUAL",
              "AUTOMATIC"
            ]
          },
          "seats": {
            "type": "number",
            "example": 9
          },
          "ownerId": {
            "type": "string",
            "description": "ID du propriétaire"
          },
          "carculVehicleId": {
            "type": "string",
            "description": "ID Carcul GPS"
          }
        },
        "required": [
          "plateNumber",
          "brand",
          "model"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string"
          },
          "fuelType": {
            "type": "string"
          },
          "transmission": {
            "type": "string"
          },
          "seats": {
            "type": "number"
          },
          "carculVehicleId": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          }
        }
      },
      "AssignManagerDto": {
        "type": "object",
        "properties": {
          "managerId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "managerId"
        ]
      },
      "UpdateVehicleStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "ASSIGNED",
              "IN_SERVICE",
              "IMMOBILIZED",
              "IN_REPAIR",
              "ACCIDENTED",
              "PENDING_INSPECTION",
              "REPOSSESSED",
              "OUT_OF_SERVICE",
              "SOLD"
            ],
            "description": "Nouveau statut du véhicule"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}